// packages/logic/src/converters/types.ts
export interface RawGeneral {
  no: number;
  owner: number;
  npcMsg: string;
  npc: number;
  npcOrg: number;
  affinity: number;
  bornYear: number;
  deadYear: number;
  newMsg: number;
  picture: string;
  imgSvr: number;
  name: string;
  ownerName: string | null;
  nationId: number;
  cityId: number;
  troopId: number;
  leadership: number;
  leadershipExp: number;
  strength: number;
  strengthExp: number;
  intel: number;
  intelExp: number;
  injury: number;
  experience: number;
  dedication: number;
  dex1: number;
  dex2: number;
  dex3: number;
  dex4: number;
  dex5: number;
  officerLevel: number;
  officerCity: number;
  permission: string;
  gold: number;
  rice: number;
  crew: number;
  crewType: number;
  train: number;
  atmos: number;
  weapon: string;
  book: string;
  horse: string;
  item: string;
  turnTime: Date;
  recentWar: Date | null;
  makeLimit: number;
  killTurn: number | null;
  block: number;
  dedLevel: number;
  expLevel: number;
  age: number;
  startAge: number;
  belong: number;
  betray: number;
  personal: string;
  special: string;
  specAge: number;
  special2: string;
  specAge2: number;
  defenceTrain: number;
  tnmt: number;
  myset: number;
  tournament: number;
  newVote: number;
  lastTurn: unknown;
  aux: unknown;
  penalty: unknown;
}

export interface RawNation {
  nation: number;
  name: string;
  color: string;
  capital: number;
  capset: number;
  gennum: number;
  gold: number;
  rice: number;
  bill: number;
  rate: number;
  rateTmp: number;
  secretLimit: number;
  chiefSet: number;
  scout: number;
  war: number;
  strategicCmdLimit: number;
  surLimit: number;
  tech: number;
  power: number;
  level: number;
  type: string;
  spy: unknown;
  aux: unknown;
}

export interface RawCity {
  city: number;
  name: string;
  level: number;
  nationId: number;
  supply: number;
  front: number;
  pop: number;
  popMax: number;
  dead: number;
  agri: number;
  agriMax: number;
  comm: number;
  commMax: number;
  secu: number;
  secuMax: number;
  trust: number;
  trade: number | null;
  def: number;
  defMax: number;
  wall: number;
  wallMax: number;
  officerSet: number;
  state: number;
  region: number;
  term: number;
  conflict: unknown;
}